requirements.txt
google-generativeai
numpy
scikit-learn
PyMuPDF
crewai
crewai-tools
python-dotenv

